Move DIS nuclear models from Physics/NuclearState into their own new area Physics/DeepInelastic/NuclearModel. This change is intended to prevent confusion between code that manages initial-state nucleon kinematic distributions (stored in Physics/NuclearState) and code that provides nuclear corrections to DIS cross sections (stored in Physics/DeepInelastic/NuclearModel).
