+++ src/Physics/DeepInelastic/NuclearModel/JGDISNuclearModel.cxx
@@ -2,7 +2,7 @@
 /*
  Copyright (c) 2003-2025, The GENIE Collaboration
  For the full text of the license visit http://copyright.genie-mc.org
- 
+
 
  J. Tena Vidal <julia.tena-vidal@ific.uv>
  Universitat de Valencia
@@ -12,7 +12,7 @@
 */
 //____________________________________________________________________________
 
-#include "Physics/NuclearState/JGDISNuclearModel.h"
+#include "Physics/DeepInelastic/NuclearModel/JGDISNuclearModel.h"
 
 using std::ostringstream;
 using namespace genie;
@@ -30,9 +30,9 @@ JGDISNuclearModel::JGDISNuclearModel(string config) :
 //____________________________________________________________________________
 
 double JGDISNuclearModel::DISACorrection (const Interaction * interaction) const {
-  if ( !interaction ) return 0; 
+  if ( !interaction ) return 0;
   double f = 1.;
-  
+
   // Nuclear modification to Fi
   // The scaling variable can be overwritten to include corrections
 
@@ -44,7 +44,7 @@ double JGDISNuclearModel::DISACorrection (const Interaction * interaction) const
   double x = kinematics.x();
   int    A = tgt.A();
 
-  if ( A >= 2 ) { 
+  if ( A >= 2 ) {
     double xv     = TMath::Min(0.65, TMath::Max(0.05, x));
     double xv2    = xv  * xv;
     double xv3    = xv2 * xv;
@@ -53,12 +53,12 @@ double JGDISNuclearModel::DISACorrection (const Interaction * interaction) const
     double xv6    = xv5 * xv;
     double xv7    = xv6 * xv;
     double xv8    = xv7 * xv;
-    double alpha = fAa0 + fAa1 * xv + fAa2 * xv2 + fAa3 * xv3 + fAa4 * xv4 + fAa5 * xv5 + fAa6 * xv6 + fAa7 * xv7 + fAa8 * xv8; 
+    double alpha = fAa0 + fAa1 * xv + fAa2 * xv2 + fAa3 * xv3 + fAa4 * xv4 + fAa5 * xv5 + fAa6 * xv6 + fAa7 * xv7 + fAa8 * xv8;
     double lnC   = fAc0 + fAc1 * TMath::Log(xv) + fAc2 * pow( TMath::Log(xv),2 );
     double C = TMath::Exp( lnC ) ;
     f *= C * pow( A, alpha ) ;
   }
-  
+
 #ifdef __GENIE_LOW_LEVEL_MESG_ENABLED__
   LOG("DISSF", pDEBUG) << "J. Gomez et. al. Nuclear factor for x of " << x << "  = " << f;
 #endif
