+++ src/Physics/DeepInelastic/NuclearModel/LinkDef.h
@@ -1,0 +1,13 @@
+#ifdef __CINT__
+
+#pragma link off all globals;
+#pragma link off all classes;
+#pragma link off all functions;
+
+#pragma link C++ namespace genie;
+
+#pragma link C++ class genie::DISNuclearModelI;
+#pragma link C++ class genie::BY00DISNuclearModel;
+#pragma link C++ class genie::BY21DISNuclearModel;
+#pragma link C++ class genie::JGDISNuclearModel;
+#endif
