+++ src/Physics/NuclearState/LinkDef.h
@@ -25,7 +25,4 @@
 #pragma link C++ class genie::SRCNuclearRecoil;
 #pragma link C++ class genie::SecondNucleonEmissionI;
 #pragma link C++ class genie::SpectralFunction2p2h;
-#pragma link C++ class genie::BY00DISNuclearModel;
-#pragma link C++ class genie::BY21DISNuclearModel;
-#pragma link C++ class genie::JGDISNuclearModel;
 #endif
